services/ops-bot/requirements.txt
python-telegram-bot==20.*
psutil